decoupled conv  from qwen specific modules
